Substantial cleanup.
Forked the repo, then did preliminary clean-up so that the system will build cleanly.
Recorded change in authorship in the ASD file.
Split the templates out to avoid undefined-function warnings.
Fixed up unused arguments, etc.
Tidied up the package definition so it would be OK with ACL's case-sensitive lisp, if we ever want to use that.
Fixed typos.