Move common functions to cl-swagger/utils.
Currently there are functions used both in the code generator and in the code generatED, but they were defined twice, once as a template and once as code definitions.
Now those functions are defined once, in cl-swagger/utils (system name) and cl-swagger.utils (package name).
Now any generated client *must* be in a package that uses cl-swagger.utils.